fix: some errors in chunk format and off by one
In the chunk format there are two different fields that were previously
tested and were always present but upon further testing may not be
present. To solve this, they have been converted to be `Option<T>`.

Additionally, there was an off by one error when calculating the length
in bytes of a chunk's compressed data. This causes errors where the
read_exact will fail. However, because chunks are padded to multiples of
4096 bytes, the read_exact will only fail if this is the last chunk of a
region file and happens to have a length that is a multiple of 4096
bytes. This is quite rare in practice but has now been fixed.